+++ CHANGELOG.md
@@ -3,10 +3,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+
 ## [Unreleased] - YYYY-MM-DD
 #### Added
 #### Changed
 - Locking `react` to latest version (finding path to `react` folder hierarchically).
+- Saving `stats.json` object with build output.
+
 #### Deprecated
 #### Removed
 #### Fixed